backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib
bcrypt==4.0.1
python-multipart
python-dotenv
alembic
